+++ NEWS.md
@@ -10,6 +10,7 @@ If you were using `google_analytics()` before to fetch the v3 API, this is now a
 * Changed `google_analytics` to be the v4 API, `google_analytics_3` now supports v3
 * change default cache token name from `.httr-oauth` to `ga.oauth` to avoid clashes
 * Change the rows fetched per page in v4 from 10,000 to 100,000 
+* Add unsampled report downloads (#44 - many thanks to @j450h1 for this work on this) 
 
 # googleAnalytics 0.4.2